Check checksum of opkg information
:Release Notes:
Check checksum of opkg information

:Detailed Notes:
At the first startup, a checksum value is generated based on the opkg
information and stored in the "/var/luna/preferences/opkg_checksum" file.
When a crash event occurs, the checksum value is compared
to check whether the official image is correct.

:Testing Performed:
Local Test : OK

:QA Notes:

:Issues Addressed:
[WRN-3755] Check modified official image

Change-Id: Ibb9ad4929233da9b44991107821173745732594d
Reviewed-on: http://gpro.lge.com/c/webosose/webos-fluentbit-plugins/+/314423
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Sukil Hong <sukil.hong@lge.com>
Tested-by: Sukil Hong <sukil.hong@lge.com>