Improve playback; events can be played if the device has the same capa.
:Release Notes:
Improve playback; events can be played if the device has the same capa.

:Detailed Notes:
Previously, when playing an input event, the event was played only with
the same device name (/dev/input/eventX) that was recorded.
The problem is, when connecting a keyboard, the device name is changed
depending on whether the keyboard is connected before booting or
the keyboard is connected after booting.
So, TCs that require a keyboard connection do not work normally.
This commit saves the list of devices and their capa. together in
the recording file. And at the time of playback (even if the device
name is different), playback can be performed on the same device or
device with the same capa.

:Testing Performed:
Local test : OK

:QA Notes:
N/A

:Issues Addressed:
[WRO-2264] Support keyboard input in input_simulator

Change-Id: I8b463c62aabc1b401e5391def4c4a69fd9af3eca
Reviewed-on: http://gpro.lge.com/c/webosose/webos-fluentbit-plugins/+/328708
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Myungchul Kim <myungchul.kim@lge.com>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>
Package-Only-Build: Myungchul Kim <myungchul.kim@lge.com>
Tested-by: Myungchul Kim <myungchul.kim@lge.com>