Modify find_issue_with_summary function to return only issues with an exact summary.
:Release Notes:
Modify find_issue_with_summary function to return only issues with an exact summary.

:Detailed Notes:
Because Jira's summary search is a 'contains (~)' search, so it gives unintended results.
That is, if we search for "\\[TAS Failed\\]\\[OSE\\]\\ ABC" as the summary, we get both of the followings.
1. [TAS Failed][OSE] ABC
2. [TAS Failed][OSE Emulator] ABC
So, after searching, compare the summary one more time to see if they match exactly.

:Testing Performed:
Local Test: OK

:QA Notes:

:Issues Addressed:
[WRP-]

Change-Id: I00aed18a25559c99b4e13865ddc400383d83355f
Reviewed-on: http://gpro.lge.com/c/webosose/webos-fluentbit-plugins/+/356968
Reviewed-by: Myungchul Kim <myungchul.kim@lge.com>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>
Tested-by: Myungchul Kim <myungchul.kim@lge.com>